fix(omg-cli): make checkVersion async, and move the message at the end of the execution
